+++ weblate/addons/models.py
@@ -124,6 +124,7 @@ class AddonsConf(AppConf):
         'weblate.addons.gettext.GettextCustomizeAddon',
         'weblate.addons.gettext.GettextAuthorComments',
         'weblate.addons.cleanup.CleanupAddon',
+        'weblate.addons.consistency.LangaugeConsistencyAddon',
         'weblate.addons.discovery.DiscoveryAddon',
         'weblate.addons.flags.SourceEditAddon',
         'weblate.addons.flags.TargetEditAddon',